azure-functions/requirements.txt
azure-functions
azure-functions-durable
azure-cosmos
azure-ai-formrecognizer
azure-storage-blob
azure-identity
